formatting

+++ src/op_brightness_contrast.cpp
@@ -21,13 +21,14 @@ THE SOFTWARE.
 */
 #include "op_brightness_contrast.hpp"
 
+#include <hip/hip_runtime.h>
+
 #include <functional>
 
-#include <hip/hip_runtime.h>
-#include "core/wrappers/image_wrapper.hpp"
 #include "common/validation_helpers.hpp"
 #include "core/detail/casting.hpp"
 #include "core/detail/type_traits.hpp"
+#include "core/wrappers/image_wrapper.hpp"
 #include "kernels/device/brightness_contrast_device.hpp"
 #include "kernels/host/brightness_contrast_host.hpp"
 
@@ -35,14 +36,13 @@ namespace roccv {
 
 template <typename BCWrappers, typename SRC_DT, typename DST_DT, int NC>
 void dispatch_brightness_contrast_channels(hipStream_t stream, const Tensor &input, const Tensor &output,
-                                            const BCWrappers &bc_wrappers, eDeviceType device) {
-    
+                                           const BCWrappers &bc_wrappers, eDeviceType device) {
     using SRC_DT_NC = detail::MakeType<SRC_DT, NC>;
     using DST_DT_NC = detail::MakeType<DST_DT, NC>;
 
     ImageWrapper<SRC_DT_NC> inputWrapper(input);
     ImageWrapper<DST_DT_NC> outputWrapper(output);
-    
+
     // Launch CPU/GPU kernel depending on requested device type.
     switch (device) {
         case eDeviceType::GPU: {
@@ -61,10 +61,9 @@ void dispatch_brightness_contrast_channels(hipStream_t stream, const Tensor &inp
 
 template <typename BCWrappers, typename SRC_DT, typename DST_DT>
 void dispatch_brightness_contrast_output_dtype(hipStream_t stream, const Tensor &input, const Tensor &output,
-                                       const BCWrappers &bc_wrappers, eDeviceType device) {
-
+                                               const BCWrappers &bc_wrappers, eDeviceType device) {
     int64_t channels = output.shape(output.layout().channels_index());
-    // 4) Select kernel dispatcher based on number of channels.
+    // Select kernel dispatcher based on number of channels.
     // clang-format off
     static const std::array<std::function<void(hipStream_t, const Tensor &, const Tensor &, const BCWrappers &, eDeviceType)>, 4>
         funcs = {dispatch_brightness_contrast_channels<BCWrappers, SRC_DT, DST_DT, 1>, dispatch_brightness_contrast_channels<BCWrappers, SRC_DT, DST_DT, 2>, dispatch_brightness_contrast_channels<BCWrappers, SRC_DT, DST_DT, 3>, dispatch_brightness_contrast_channels<BCWrappers, SRC_DT, DST_DT, 4>};
@@ -77,68 +76,60 @@ void dispatch_brightness_contrast_output_dtype(hipStream_t stream, const Tensor 
 
 template <typename BCWrappers, typename SRC_DT>
 void dispatch_brightness_contrast_input_dtype(hipStream_t stream, const Tensor &input, const Tensor &output,
-                                            const BCWrappers &bc_wrappers, eDeviceType device) {
-    
+                                              const BCWrappers &bc_wrappers, eDeviceType device) {
     eDataType output_dtype = output.dtype().etype();
-    
-    // 3) Select kernel dispatcher based on a base output datatype.
+
+    // Select kernel dispatcher based on a base output datatype.
     // clang-format off
     static const std::unordered_map<eDataType, std::function<void(hipStream_t, const Tensor &, const Tensor &, const BCWrappers &, eDeviceType)>>
         funcs = {
-            {eDataType::DATA_TYPE_U8, dispatch_brightness_contrast_output_dtype<BCWrappers, SRC_DT, uchar>},
-            {eDataType::DATA_TYPE_U16,  dispatch_brightness_contrast_output_dtype<BCWrappers, SRC_DT, ushort>},
-            {eDataType::DATA_TYPE_S16,  dispatch_brightness_contrast_output_dtype<BCWrappers, SRC_DT, short>},
-            {eDataType::DATA_TYPE_S32,  dispatch_brightness_contrast_output_dtype<BCWrappers, SRC_DT, int>},
+            {eDataType::DATA_TYPE_U8,  dispatch_brightness_contrast_output_dtype<BCWrappers, SRC_DT, uchar>},
+            {eDataType::DATA_TYPE_U16, dispatch_brightness_contrast_output_dtype<BCWrappers, SRC_DT, ushort>},
+            {eDataType::DATA_TYPE_S16, dispatch_brightness_contrast_output_dtype<BCWrappers, SRC_DT, short>},
+            {eDataType::DATA_TYPE_S32, dispatch_brightness_contrast_output_dtype<BCWrappers, SRC_DT, int>},
             {eDataType::DATA_TYPE_F32, dispatch_brightness_contrast_output_dtype<BCWrappers, SRC_DT, float>},
         };
     // clang-format on
     auto func = funcs.at(output_dtype);
     if (func == 0) throw Exception("Not mapped to a defined function.", eStatusType::INVALID_OPERATION);
     func(stream, input, output, bc_wrappers, device);
-
 }
 
 template <typename BCWrappers>
 void dispatch_brightness_contrast_bc_dtype(hipStream_t stream, const Tensor &input, const Tensor &output,
-                                        const BCWrappers &bc_wrappers, eDeviceType device) {
-    
+                                           const BCWrappers &bc_wrappers, eDeviceType device) {
     eDataType input_dtype = input.dtype().etype();
 
-    // 2) Select kernel dispatcher based on a base input datatype.
+    // Select kernel dispatcher based on a base input datatype.
     // clang-format off
     static const std::unordered_map<eDataType, std::function<void(hipStream_t, const Tensor &, const Tensor &, const BCWrappers &, eDeviceType)>>
         funcs = {
-            {eDataType::DATA_TYPE_U8, dispatch_brightness_contrast_input_dtype<BCWrappers, uchar>},
-            {eDataType::DATA_TYPE_U16,  dispatch_brightness_contrast_input_dtype<BCWrappers, ushort>},
-            {eDataType::DATA_TYPE_S16,  dispatch_brightness_contrast_input_dtype<BCWrappers, short>},
-            {eDataType::DATA_TYPE_S32,  dispatch_brightness_contrast_input_dtype<BCWrappers, int>},
+            {eDataType::DATA_TYPE_U8,  dispatch_brightness_contrast_input_dtype<BCWrappers, uchar>},
+            {eDataType::DATA_TYPE_U16, dispatch_brightness_contrast_input_dtype<BCWrappers, ushort>},
+            {eDataType::DATA_TYPE_S16, dispatch_brightness_contrast_input_dtype<BCWrappers, short>},
+            {eDataType::DATA_TYPE_S32, dispatch_brightness_contrast_input_dtype<BCWrappers, int>},
             {eDataType::DATA_TYPE_F32, dispatch_brightness_contrast_input_dtype<BCWrappers, float>},
         };
     // clang-format on
     auto func = funcs.at(input_dtype);
     if (func == 0) throw Exception("Not mapped to a defined function.", eStatusType::INVALID_OPERATION);
     func(stream, input, output, bc_wrappers, device);
-
 }
 
 void BrightnessContrast::operator()(hipStream_t stream, const roccv::Tensor &input, const roccv::Tensor &output,
-                    std::optional<std::reference_wrapper<const Tensor>> brightness, 
-                    std::optional<std::reference_wrapper<const Tensor>> contrast,
-                    std::optional<std::reference_wrapper<const Tensor>> brightnessShift,
-                    std::optional<std::reference_wrapper<const Tensor>> contrastCenter,
-                    eDeviceType device) const {
-    
+                                    std::optional<std::reference_wrapper<const Tensor>> brightness,
+                                    std::optional<std::reference_wrapper<const Tensor>> contrast,
+                                    std::optional<std::reference_wrapper<const Tensor>> brightnessShift,
+                                    std::optional<std::reference_wrapper<const Tensor>> contrastCenter,
+                                    eDeviceType device) const {
     // Validate input tensor
     CHECK_TENSOR_DEVICE(input, device);
-    CHECK_TENSOR_DATATYPES(input, DATA_TYPE_U8, DATA_TYPE_U16, DATA_TYPE_S16,
-                           DATA_TYPE_S32, DATA_TYPE_F32);
-    // TODO: copied over from convert to, verify these layouts/channels are the desired ones
+    CHECK_TENSOR_DATATYPES(input, DATA_TYPE_U8, DATA_TYPE_U16, DATA_TYPE_S16, DATA_TYPE_S32, DATA_TYPE_F32);
     CHECK_TENSOR_LAYOUT(input, TENSOR_LAYOUT_HWC, TENSOR_LAYOUT_NHWC);
     CHECK_TENSOR_CHANNELS(input, 1, 2, 3, 4);
 
     // Validate output tensor
-    CHECK_TENSOR_DATATYPES(output, DATA_TYPE_U8, DATA_TYPE_U16, DATA_TYPE_S16,
-                           DATA_TYPE_S32, DATA_TYPE_F32);
+    CHECK_TENSOR_DATATYPES(output, DATA_TYPE_U8, DATA_TYPE_U16, DATA_TYPE_S16, DATA_TYPE_S32, DATA_TYPE_F32);
     CHECK_TENSOR_COMPARISON(input.device() == output.device());
     CHECK_TENSOR_COMPARISON(input.shape() == output.shape());
 
@@ -146,17 +137,19 @@ void BrightnessContrast::operator()(hipStream_t stream, const roccv::Tensor &inp
     eDataType input_dtype = input.dtype().etype();
     int64_t input_batch = input.shape(input.layout().batch_index());
     eDataType output_dtype = output.dtype().etype();
-    eDataType bc_dtype = (input_dtype == eDataType::DATA_TYPE_S32 || output_dtype == eDataType::DATA_TYPE_S32) ? eDataType::DATA_TYPE_F64 : eDataType::DATA_TYPE_F32; 
+    eDataType bc_dtype = (input_dtype == eDataType::DATA_TYPE_S32 || output_dtype == eDataType::DATA_TYPE_S32)
+                             ? eDataType::DATA_TYPE_F64
+                             : eDataType::DATA_TYPE_F32;
 
-    auto validate_bc_param = [&](const auto& param_opt) {
-      if (param_opt.has_value()) {
-        const Tensor& param = param_opt->get();
-        CHECK_TENSOR_COMPARISON(param.dtype().etype() == bc_dtype);
-        CHECK_TENSOR_LAYOUT(param, TENSOR_LAYOUT_N);
-        CHECK_TENSOR_COMPARISON(param.shape(param.layout().batch_index()) == 1 ||
-                                param.shape(param.layout().batch_index()) == input_batch);
-        CHECK_TENSOR_COMPARISON(input.device() == param.device());
-      }
+    auto validate_bc_param = [&](const auto &param_opt) {
+        if (param_opt.has_value()) {
+            const Tensor &param = param_opt->get();
+            CHECK_TENSOR_COMPARISON(param.dtype().etype() == bc_dtype);
+            CHECK_TENSOR_LAYOUT(param, TENSOR_LAYOUT_N);
+            CHECK_TENSOR_COMPARISON(param.shape(param.layout().batch_index()) == 1 ||
+                                    param.shape(param.layout().batch_index()) == input_batch);
+            CHECK_TENSOR_COMPARISON(input.device() == param.device());
+        }
     };
 
     validate_bc_param(brightness);
@@ -166,35 +159,34 @@ void BrightnessContrast::operator()(hipStream_t stream, const roccv::Tensor &inp
 
     auto compute_cc_default = [&]() -> double {
         switch (input_dtype) {
-            case eDataType::DATA_TYPE_U8:  return 1u << (8 - 1);
-            case eDataType::DATA_TYPE_U16: return 1u << (16 - 1);
-            case eDataType::DATA_TYPE_S16: return 1u << (16 - 2);
-            case eDataType::DATA_TYPE_S32: return 1u << (32 - 2);
-            case eDataType::DATA_TYPE_F32: return 0.5;
-            default: return 0.5;
+            case eDataType::DATA_TYPE_U8:
+                return 1u << (8 - 1);
+            case eDataType::DATA_TYPE_U16:
+                return 1u << (16 - 1);
+            case eDataType::DATA_TYPE_S16:
+                return 1u << (16 - 2);
+            case eDataType::DATA_TYPE_S32:
+                return 1u << (32 - 2);
+            case eDataType::DATA_TYPE_F32:
+                return 0.5;
+            default:
+                return 0.5;
         }
     };
 
-    // 1) Select kernel dispatcher based on BC datatype
+    // Select kernel dispatcher based on BC datatype
     if (bc_dtype == eDataType::DATA_TYPE_F32) {
-        GroupBCWrappers<float> wrappers {
-            BCWrapper<float>(brightness, 1.0f),
-            BCWrapper<float>(contrast, 1.0f),
-            BCWrapper<float>(brightnessShift, 0.0f),
-            BCWrapper<float>(contrastCenter, compute_cc_default())
-        };
+        GroupBCWrappers<float> wrappers{BCWrapper<float>(brightness, 1.0f), BCWrapper<float>(contrast, 1.0f),
+                                        BCWrapper<float>(brightnessShift, 0.0f),
+                                        BCWrapper<float>(contrastCenter, static_cast<float>(compute_cc_default()))};
         dispatch_brightness_contrast_bc_dtype<GroupBCWrappers<float>>(stream, input, output, wrappers, device);
     } else if (bc_dtype == eDataType::DATA_TYPE_F64) {
-        GroupBCWrappers<double> wrappers {
-            BCWrapper<double>(brightness, 1.0),
-            BCWrapper<double>(contrast, 1.0),
-            BCWrapper<double>(brightnessShift, 0.0),
-            BCWrapper<double>(contrastCenter, compute_cc_default())
-        };
+        GroupBCWrappers<double> wrappers{BCWrapper<double>(brightness, 1.0), BCWrapper<double>(contrast, 1.0),
+                                         BCWrapper<double>(brightnessShift, 0.0),
+                                         BCWrapper<double>(contrastCenter, compute_cc_default())};
         dispatch_brightness_contrast_bc_dtype<GroupBCWrappers<double>>(stream, input, output, wrappers, device);
     } else {
         throw Exception("Not mapped to a defined function.", eStatusType::INVALID_OPERATION);
     }
-
 }
-}   // namespace roccv
+}  // namespace roccv
